WKB type inference and refactored capacity (#299)
In order to parse an arbitrary WKB array to a geoarrow array we need to
know which geometry type it represents (though in the future we'll be
able to alternatively first parse to a MixedGeometryArray and then
downcast). This PR starts this process.

Originally I implement a simple struct containing booleans, which would
eliminate type possibilities until only a single type is available.

But then I realized that this is making an extra pass through the data,
and there are still four more passes (one to collect `WKB` objects; one
to downcast to _typed_ WKB objects, e.g. `WKBPoint`; one to count
capacities; and the last to fill arrays. The first two should be very
fast. Ideally they could be coalesced into one but I got lifetime errors
when I tried that).

At that point, I switched to count array capacities in the same original
pass as inferring the WKB type. This should remove at least one pass
through the data.

### Change list

- New top-level `parse_to_geoarrow` method on the WKB Array that parses
to a zero-copy geoarrow array.
- Add new `Capacity` structs, e.g. `MultiPolygonCapacity` that hold the
relevant capacities that make up the array.
- Change existing methods `with_capacities` and `reserve` to take this
struct instead of 1-4 individual numbers.
- Remove `count_from_iter` in favor of e.g.
`PolygonCapacity::from_polygons`
- Add `from_wkb` method to construct a builder from a slice of `WKB`
objects. This also allows to set the coord type.

TODO: 

- [x] Remove original type inference before I moved to count coordinate
numbers in the same pass.
- Infer the output array type from the WKB statistics
- Connect that type to creating arrays with the given capacity and
filling them. You should be able to use `Builder::with_capacity` and
`extend_from_iter` without any extra allocations.
- [x] Add single-* variants onto the capacity counters. E.g. the
`MultiPolygonCapacity` should have an `add_polygon` method to add a
simple polygon
- [x] Add `add_rect` to `PolygonCapacity` and `PolygonBuilder`.
- [ ] Next PR: capacity structs should also hold a null count and
instantiate the null buffer to the known length if null count is > 0.
This should also make it possible to accurately add a polygon capacity
to a multipolygon capacity and so on.
https://github.com/geoarrow/geoarrow-rs/issues/301