Handle timestamps in `read_postgres` (#420)
This is a relative hack around the geozero type system because we have
an already-parsed
datetime value (as a chrono type) and geozero only supports
string-formatted timestamps.

At some point it will be good to refactor the table builder to make the
coupling with geozero not quite as tight


~~This needs a bit of thought behind it. In particular, sqlx exposes
timestamps as chrono objects, while geozero exposes them as iso8601
strings. I don't want to serialize the chrono objects to _string_ and
then decode them _again_. Which means I need to uncouple a bit from
geozero. The table builder needs to be generalized to _arrow_ and have
an optional geozero api, but also have a direct postgres -> arrow API~~