Vendor object-store-python (#572)
For https://github.com/geoarrow/geoarrow-rs/issues/551. Closes
https://github.com/geoarrow/geoarrow-rs/issues/552

object-store-python is not currently actively maintained (or at least as
much as I'd like). I currently have 3 open PRs and a few open issues. I
want to be able to move a bit quicker, so I'm vendoring some relevant
code in this repo for now.