Parse Parquet geometry column after reading all files (#614)
This is an alternative solution to #609.

One thing we always have to be careful about is ensuring that the schema
across record batches is always the same. When the GeoParquet metadata
declares the geometry type of the underlying data, we always know what
to parse the geometry column to. But it's also allowed for the
GeoParquet metadata to not declare the column types. In this case, we
parse to a mixed array and then downcast.

But it would be bad to downcast each file of a dataset independently, as
it would be possible for one file in a dataset to contain only points
and for another to contain only polygons, even where the rest of the
schema is the same.

In this PR, we defer the parsing from WKB until after the data has
already been materialized. Maybe this will slightly increase the wall
time in the case where we have to wait a while to fetch the data and
_then_ parse the data. But it's simpler than switching on whether the
relevant metadata exists

---------

Co-authored-by: Nicholas Roberts <nicholas.roberts.au@gmail.com>