3d builders (#725)
This extends builders to support 3d coordinates. 

Note that users still have to know in advance which dimensionality of
array they want to create. So it isn't ideal for unknown-dimension
input, but still an improvement.

This also fixes GEOS interop (I was previously using the incorrect GEOS
function to see the number of dimensions)