Add `push_geometry` to `GeometryArrayBuilder` (#820)
### Change list

- Move `prefer_multi` from a parameter when adding a single geometry to
a parameter on builder initialization. Realistically, the `prefer_multi`
keyword is used when downcasting is planned, such as when using an
unknown geometry source. In these cases, we'll always know on array
creation whether we'll want to downcast later.
- This means that we can now add `push_geometry` onto the
`GeometryArrayBuilder` trait, because we have a single signature that
all builders can follow. This should make it a lot easier to manage
stuff like macros because we won't need a different impl for points,
geometries, and geometry collections.