Unknown capacity and builder (#890)
For https://github.com/geoarrow/geoarrow-rs/issues/887.

This adds a new native geometry array and builder, for now called
`UnknownGeometryArray` and `UnknownGeometryBuilder`. (Hopefully we'll
find better names for this in the future, perhaps even just
`GeometryArray`, because it's the most generic array). This is a mixed
geometry array that supports _all_ dimensions (XY, XYZ, and in the
future XYM and XYZM).

This is really useful for parsers because we often don't know the
geometry schema in advance. E.g. when parsing WKT or WKB from unknown
origin, we don't know anything about the data. But this array type still
gives us a totally static schema, so we can do streaming stuff.