Remove owned slice (#891)
Closes #889 

When I was originally implementing scalars, I didn't really understand
how cheap `Arc`s are. I was thinking that for scalars we don't
necessarily want to hold on to the full geometry array just to keep one
scalar around. While that's sometimes true, we also don't necessarily
want to make a full memcopy of the scalar whenever we access it.
Sometimes it's better to just clone the `Arc` and maintain the geometry
offset.

`owned_slice` was an implementation of this full memcopy for the given
offset and length. It was never fully implemented and at this point it's
better to just remove it.