downcast geometry array (#918)
todo:

- [x] Change downcast to be implemented in terms of `resolve_types` and
`cast`.
- [x] Handle slicing in `GeometryArray::has_points`. I.e. bring back the
slice offset and length. Then if the array has been sliced and the point
array exists, check if the point values are within the bounds of the
current slice. The fastest way to do this should be to check the length
of the point array, then as you're iterating through the type_ids array,
once you've seen that many ids pointing to the point array, you know the
array is fully there. Or, in the cast of downcasting you really only
care about whether _any_ geometry exists. So you can short-circuit as
soon as you've seen the first point.
- [x] Ensure we apply slicing when exporting the GeometryArray. E.g. in
the `TryFrom` impl from GeometryArray to `PointArray`, we need to call
the slice when we export.
- [ ] recursively expand the types from the geometry collection array.
But only if the geometry collection array has only a single geometry per
row and it could be split up.

Closes https://github.com/geoarrow/geoarrow-rs/issues/416