GeometryStreamBuilder (#965)
### Change list

- Instead of using really complex code to figure out which underlying
array to push each individual `xy` call to, we use an underlying
`GeoWriter` to collect each geometry into a scratch buffer, and then
push that geometry using our existing `push_geometry` APIs. This
simplifies the code a _lot_.
- It may have a negative performance impact because we're allocating
extra heap buffers for the small vecs. In the future, if we want to
optimize this, we may want to create a version of `geo-types` that uses
something like bumpalo

Closes https://github.com/geoarrow/geoarrow-rs/issues/365