feat(io): Convert FlatGeobuf reader to RecordBatchReader iterator (#933)
This PR updates the FlatGeobuf reader (the sync version only for now) to
expose data as an iterator. This means that applications like Python can
access data larger than memory, and process it in batches.

- [x] This was blocked on Flatgeobuf exposing the `Seekable` and
`NotSeekable` marker structs, but that happened in
https://github.com/flatgeobuf/flatgeobuf/pull/402

Closes https://github.com/geoarrow/geoarrow-rs/issues/932

For https://github.com/geoarrow/geoarrow-rs/issues/594, for
https://github.com/geoarrow/geoarrow-rs/issues/596