Simplify builder apis (#1021)
### Change list

#### Simplify builder constructors

So instead of both `new` and `new_with_options`

```rs
// OLD
impl LineStringBuilder {
	/// Creates a new empty [`LineStringBuilder`].
	pub fn new(dim: Dimension) -> Self;

	/// Creates a new empty [`LineStringBuilder`] with the provided options.
	pub fn new_with_options(
		dim: Dimension,
		coord_type: CoordType,
		metadata: Arc<Metadata>,
	) -> Self;
```

we now just have a `new` that takes in a relevant type

```rs
// NEW
impl LineStringBuilder {
	/// Creates a new empty [`LineStringBuilder`].
	pub fn new(typ: LineStringType) -> Self;
```

And similarly, instead of having with `with_capacity` and
`with_capacity_with_options`, we now have just `with_capacity`.

This also simplifies the signature of APIs like

```rs
	pub fn from_line_strings(
		geoms: &[impl LineStringTrait<T = f64>],
		dim: Dimension,
		coord_type: CoordType,
		metadata: Arc<Metadata>,
	) -> Self;
``` 

which becomes

```rs
	pub fn from_line_strings(geoms: &[impl LineStringTrait<T = f64>], typ: LineStringType) -> Self;
```

#### Convert `From` impls that used to be in terms of `Dimension`, like

```rs
impl TryFrom<(&dyn Array, Dimension)> for LineStringArray
```

to be in terms of the relevant type instead

```rs
impl TryFrom<(&dyn Array, LineStringType)> for LineStringArray {
```

This ensures that the correct coord type and metadata is propagated, and
we don't have to use `Default::default` on either of those.

- Remove `From` impls like

```rs
impl<G: LineStringTrait<T = f64>> From<(&[G], Dimension)> for LineStringBuilder {
```

and

```rs
impl<G: LineStringTrait<T = f64>> From<(Vec<Option<G>>, Dimension)> for LineStringBuilder {
```

For one, it was super awkward that in the non-nullable case we accept a
slice but in the nullable case we only accepted a vec. We can't have
both in terms of slices because that gives an overlapping trait error.
Because G could be implemented in the future for `Option<G>` and Rust
wouldn't know which trait to choose.

#### Remove `From<Builder> for Array`. Instead, have only the named
method `finish()` which converts a builder to an array.

#### Remove conversions between e.g. `MultiLineStringArray` and
`PolygonArray`, which have the same physical layout but different
interpretations.

#### Better and more concise handling of field metadata using the new
upstream extension APIs


```rs
// OLD
   fn try_from((arr, field): (&dyn Array, &Field)) -> Result<Self> {
       let geom_type = NativeType::try_from(field)?;
       let dim = geom_type
           .dimension()
           .ok_or(GeoArrowError::General("Expected dimension".to_string()))?;
       let mut arr: Self = (arr, dim).try_into()?;
       let metadata = Arc::new(Metadata::try_from(field)?);
       arr.data_type = arr.data_type.clone().with_metadata(metadata);
       Ok(arr)
   }
```

```rs
// NEW
   fn try_from((arr, field): (&dyn Array, &Field)) -> Result<Self> {
       let typ = field.try_extension_type::<GeometryCollectionType>()?;
       (arr, typ).try_into()
   }
```