Add `downcast_geometry_array` macro (#1031)
I'm tired today and accidentally pushed to main in
https://github.com/geoarrow/geoarrow-rs/commit/80e905b1086deb357ca69819b5740ebacc269714.
(github branch protection settings are so weird. I can't push with
`--force`, but I _can_ push without `--force`. I since turned on "do not
allow bypassing these settings", which might help prevent this again in
the future)

Then https://github.com/geoarrow/geoarrow-rs/pull/1030 auto-merged on
top of it and I decided it wasn't worth the work to rewrite the history.