Restore geozero integration into geoarrow-array (#1036)
### Change list

- Adds optional geozero dependency via feature flag in `geoarrow-array`.
It's simpler to have this in the same module than a separate
`geoarrow-geozero`, because the traits are in `geozero` and the structs
are defined in `geoarrow-array`. So the orphan rule would say there
would need to be _new_ wrapper structs defined in `geoarrow-geozero`. I
think having a self-contained geozero module is fine.
- Implements exporting traits (`GeozeroGeometry`) on all array and
scalar types defined in `geoarrow-array`.
- Adds traits for importing data from geozero. This is kinda hard for
arbitrary geometries because geozero seems to overload a
`GeometryCollection` as either a top-level sequence of geometries or as
a literal GeometryCollection object.
- Adds some basic tests.