Improved null handling for `GeometryArray` (#1072)
### Change list

- Partly a legacy holdover but we have a "MixedArray" concept. This is
now used only as the child of a `GeometryCollectionArray`. So a
`GeometryCollectionArray` holds an offset buffer + a Mixed array. This
means that the `MixedArray` itself will **never** need to be concerned
with null input. Because the `GeometryCollectionArray` will always
itself handle null input before it reaches the `MixedArray`. This means
we can simplify a lot of the `MixedArrayBuilder` APIs to only take in
non-null geometries.
- Modify `GeoArrowArray` trait to have `logical_nulls` but not `nulls`.
The upstream `Array` trait has a weird separation between
[`nulls`](https://docs.rs/arrow/latest/arrow/array/trait.Array.html#tymethod.nulls)
and
[`logical_nulls`](https://docs.rs/arrow/latest/arrow/array/trait.Array.html#method.logical_nulls).
`nulls` is always a reference whereas `logical_nulls` may be computed.
In our case, we use union arrays heavily in our `GeometryArray`. In
upstream, `nulls` is **wrong** for a `UnionArray`. You have to **know**
that the result will be invalid for `UnionArray`. This is an easy way to
shoot yourself in the foot. Here, we remove the `nulls` API and only
present `logical_nulls` so that the result is always valid.
- Additionally, we specialize `is_null` and `null_count` so that those
code paths don't need to construct the `logical_nulls`.
- Simple rename of internal `NullBuffer` in each geometry array type
from `validity` to `nulls`.
- Fix pushing null geometries to GeometryCollectionBuilder
- Remove unused code from MixedGeometryBuilder
- Add logical nulls tests for `GeometryArray` and
`GeometryCollectionArray`