Fix deferred null handling in `GeometryBuilder` (#1073)
- Fix deferred null application in geometry builder. The
`GeometryBuilder` has an (internal) concept of "deferred nulls". That
is, before a valid geometry has been pushed, we don't know where to push
a null value. Ideally we want to push a null value to the same array as
future valid geometries, so that it's easier to "downcast" to a single
concrete array type if desired.
- Previously the deferred nulls were broken because we were pushing the
nulls to the children but not updating the `types` and `offsets` in the
top-level union array.
- Add a test where all nulls and no valid geometries are pushed. 
- Add a test where partial nulls are pushed.