Implement downcasting to concrete types. (#1080)
Downcasting has long been a challenge of how to implement it without too
much complexity, but I think this PR (on top of
https://github.com/geoarrow/geoarrow-rs/pull/1076) is pretty good!

- `downcast::get_type_ids`: given a sequence of arrays, find the "type
ids", e.g. native type + dimension, for the input. This returns a
`HashSet` with all type ids in the input.
- `infer_from_native_type_and_dimension`: given a `HashSet` with type
ids, find the native type + dimension combo that we can cast to.
    - If only one type, easy, we can cast to that.
    - If two types, choose the multi type if possible.
- Else, return `None`, meaning that there's no single native type we can
downcast to.
- Implement casting for multi-geometry types to single-geometry types
- Implement casting for geometry array to primitive array types.
- Fix pushing empty `MultiLineString` into `LineStringBuilder`, and same
for point and polygon
- Add `GeoArrowError::Cast` (or potentially we could use
`ArrowError::Cast`? 🤷‍♂️ )
- Implement conversion from `wkt::types::Dimension` to
`geoarrow_schema::Dimension`
- Edit accessors of `GeometryCapacity` to pass in dimension as an
argument. (the ordering of the dimensions in the internal arrays should
be an implementation detail only)

TODO: 

- [x] Add tests for casting multi-geometries to single geometries when
those multi-arrays have single or empty values
- [x] Add tests that casting multi geoms to single geoms fail with
actual multi geoms
- [x] Add tests that casting geometry array to primitive type
succeeds/fails