Implement ST_Point (#1106)
It seems that the extension metadata isn't getting passed through here
correctly.

I've added a test case that shows that there are two places internally
where the extension metadata **is correct** (assertions pass; the result
of `return_field_from_args` and the value of `args.return_field` in
`invoke_with_args`). But if I collect the data into record batches with
`sql_df.collect()`, the result schema is **missing** the Arrow extension
metadata (assertion fails).

@timsaucer do you know if I'm doing something wrong here? Or maybe this
is an edge case because I'm passing a _literal_ into the SQL command
(`"SELECT ST_Point(-71.104, 42.315);"`)?

Elsewhere the extension metadata seemed to correctly exist on the schema
of the resulting record batches. In particular, in the test testing
`ST_AsBinary` the relevant line succeeds
https://github.com/geoarrow/geoarrow-rs/blob/26f2b6a096b09857d5d03a8fbc9bb6eb22ca2cb4/rust/geodatafusion/src/udf/native/io/wkb.rs#L210-L211