Add `push_geometry` method to `GeoArrowArrayBuilder` trait (#1124)
Note that adding `push_geometry` causes this trait to no longer be
dyn-compatible, because there's a generic for `GeometryTrait` in
`push_geometry`.

This is not yet exported publicly so that we can iterate on the design,
and in particular decide whether we need dyn-compatibility or not.
Hopefully when we implement more format parsers it'll be more clear.

Closes https://github.com/geoarrow/geoarrow-rs/issues/1118