Delete the bulk of the old geoarrow crate (#1126)
As described in https://github.com/geoarrow/geoarrow-rs/pull/1097, the
old `geoarrow` crate is being refactored into a monorepo of smaller
crates. Most of the core refactoring has been done (except for IO,
primarily) and so this PR deletes a large amount of this old code that
has no more purpose in this repo.

👋