feat: Define `Arc<Metadata>` as the associated type for `ExtensionType` (#1156)
### Change list

- GeoArrow metadata can be somewhat sizeable, at least when a large
PROJJSON string is contained within it. Because of this, storing the
`Metadata` under an `Arc` makes sense so that cloning the type is very
cheap.

We had already been storing `Arc<Metadata>` inside of the GeoArrow
extension types, but we had been defining the associated type on
`ExtensionType` as `Metadata`. That means that whenever someone accessed
`ExtensionType::metadata` they would get a `&Metadata` instead of a
`&Arc<Metadata>`. The former is more expensive to clone than the latter,
and you would need to put it into a _new_ `Arc`. Now it's possible to
just clone the existing `Arc`.