fix(geoarrow-array): Fix persisting CRS when creating `GeometryArray` from `Field` (#1326)
Closes #1327 . I thought it was user error with casting (and indeed I
think we do need some more docs there), but the root error here was the
metadata on the field getting dropped accidentally when constructing the
`GeometryArray`.

This is failing in Lonboard with an array with CRS provided.

We should document that for now, cast always needs to have the same CRS
(`Metadata`) as what's already on the array.

What's the simplest way, most user-friendly way to handle this?